+++ src/app/products/products.module.ts
@@ -1,5 +1,5 @@
 import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
+import { SharedModule } from '../shared/shared.module';
 
 import { ProductComponent } from './components/product/product.component';
 import { ProductListComponent } from './components/product-list/product-list.component';
@@ -11,7 +11,7 @@ import {ProductsService} from './services/products.service';
     ProductListComponent
   ],
   imports: [
-    CommonModule
+    SharedModule
   ],
   // providers: [ProductsService],
   exports: [